Домашнее задание к лекции 10 (issue #9)
