+++ hw10-bite-code/src/main/java/ru/otus/homework/App.java
@@ -1,5 +1,6 @@
 package ru.otus.homework;
 
+import ru.otus.homework.logging.CalculableImpl;
 import ru.otus.homework.logging.Ioc;
 import ru.otus.homework.logging.Calculable;
 
@@ -9,7 +10,7 @@ public class App {
 
     public static void main(String[] args) {
         System.out.println("Application launched!");
-        Calculable calculable = Ioc.createInstanceWithLogging();
+        Calculable calculable = new Ioc<Calculable>().createInstance( new CalculableImpl(),Calculable.class);
         calculable.calculate(new Random().nextInt(0,100));
     }
