+++ hw10-bite-code/src/main/java/ru/otus/homework/logging/Ioc.java
@@ -1,54 +1,20 @@
 package ru.otus.homework.logging;
 
 import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
-import java.util.Arrays;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-public class Ioc {
+public class Ioc<I> {
 
-    public static Calculable createInstanceWithLogging() {
-        InvocationHandler handler = new LoggingInvocationHandler(new CalculableImpl());
-        return (Calculable)
+    @SuppressWarnings("unchecked")
+    public <T extends I> I createInstance(T instance, Class<I> interfaceType) {
+        InvocationHandler handler = new LoggingInvocationHandler<I>(instance);
+
+        return (I)
                 Proxy.newProxyInstance(
                         Ioc.class.getClassLoader(),
-                        new Class<?>[]{Calculable.class},
+                        new Class<?>[]{interfaceType},
                         handler
                 );
-    }
-
-    static class LoggingInvocationHandler implements InvocationHandler {
-
-        private final Calculable calculableClass;
-
-        public LoggingInvocationHandler(Calculable calculableClass) {
-            this.calculableClass = calculableClass;
-        }
-
-        @Override
-        public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-            boolean isShouldBeLogged = isShouldBeLogged(method);
-
-            if (isShouldBeLogged) {
-                String argsInline = Stream.of(args)
-                        .map(Object::toString)
-                        .collect(Collectors.joining(","));
-                System.out.println("Executed method: " + method.getName() + ", params: " + argsInline);
-            }
-
-            return method.invoke(calculableClass, args);
-        }
-
-        private boolean isShouldBeLogged(Method method) {
-            String invokeMethodName = method.getName();
-
-            return Arrays.stream(calculableClass.getClass().getDeclaredMethods())
-                    .anyMatch(declaredMethod -> declaredMethod.getName().equals(invokeMethodName) &&
-                            (declaredMethod.getDeclaredAnnotation(Log.class) != null)
-                    );
-        }
     }
 
 }
